+++ src/zsdtdx/params.py
@@ -80,6 +80,7 @@ class TDXParams:
     INFOHARBOR_BLOCK_REMOTE_FILE = "infoharbor_block.dat"
     # 股票到通达信行业码、研究行业码
     TDXHY_REMOTE_FILE = "tdxhy.cfg"
-    # 行业板块名称表打在该 zip 内，成员为 tdxzs.cfg
+    # 行业板块名称表打在该 zip 内，成员为 tdxzs.cfg；长名补表成员为 ilong.dat
     ZHB_ZIP_REMOTE_FILE = "zhb.zip"
     TDXZS_ZIP_MEMBER = "tdxzs.cfg"
+    ILONG_ZIP_MEMBER = "ilong.dat"